refactor(ai): move current date helper to utils
