+++ packages/ai/src/prompts/content-editor.ts
@@ -1,5 +1,5 @@
-import { formatCurrentDate } from "@notra/ai/prompts/current-date";
 import type { ContentEditorChatPromptParams } from "@notra/ai/types/prompts";
+import { formatCurrentDate } from "@notra/ai/utils/current-date";
 import dedent from "dedent";
 
 export function getContentEditorChatPrompt(
