+++ packages/ai/src/utils/current-date.ts
